backend - added check for bookedSeats while deleting a show

+++ server/routes/showsRoute.js
@@ -2,15 +2,16 @@ const router = require('express').Router();
 const { response } = require('express');
 const authMiddleware = require('../middlewares/authMiddleware');
 const Shows = require('../models/showsModels');
+const Bookings = require('../models/bookingModels')
 
 //add a show
-router.post('/add-show', authMiddleware, async(req,res)=>{
+router.post('/add-show', authMiddleware, async (req, res) => {
     try {
         const newShow = new Shows(req.body);
         await newShow.save();
         res.send({
-            success:true,
-            message:"Show Added"
+            success: true,
+            message: "Show Added"
         })
     } catch (err) {
         res.send({
@@ -21,29 +22,29 @@ router.post('/add-show', authMiddleware, async(req,res)=>{
 })
 
 //get all shows by theatre
-router.post('/get-all-shows-by-theatre',authMiddleware, async(req,res)=>{
-    try{
-        const shows = await Shows.find({theatre:req.body.theatreId}).populate('movie');
+router.post('/get-all-shows-by-theatre', authMiddleware, async (req, res) => {
+    try {
+        const shows = await Shows.find({ theatre: req.body.theatreId }).populate('movie');
         res.send({
-            success:true,
-            message:"Shows Fetched",
-            data:shows
+            success: true,
+            message: "Shows Fetched",
+            data: shows
         })
-    }catch(err){
+    } catch (err) {
         res.send({
-            success:false,
-            message:err.message
+            success: false,
+            message: err.message
         })
     }
 })
 
 //update shows
-router.put('/update-shows', authMiddleware, async(req,res)=>{
+router.put('/update-shows', authMiddleware, async (req, res) => {
     try {
-        await Shows.findByIdAndUpdate(req.body.showsId,req.body);
+        await Shows.findByIdAndUpdate(req.body.showsId, req.body);
         res.send({
-            success:true,
-            message:"shows Updated"
+            success: true,
+            message: "shows Updated"
         })
     } catch (err) {
         res.send({
@@ -54,13 +55,23 @@ router.put('/update-shows', authMiddleware, async(req,res)=>{
 })
 
 //delete shows
-router.delete('/delete-shows', authMiddleware, async (req,res)=>{
+router.delete('/delete-shows', authMiddleware, async (req, res) => {
     try {
-        await Shows.deleteOne({ "_id" : req.body.showId });
-        res.send({
-            success: true,
-            message: "Show Deleted"
-        })
+        const currentShow = await Shows.findById(req.body.showId);
+        if (currentShow.bookedSeats.length > 0) {
+            res.send({
+                success: false,
+                message: "Cannot delete show with active bookings"
+            })
+        }
+        else {
+            await Shows.deleteOne({ "_id": req.body.showId });
+
+            res.send({
+                success: true,
+                message: "Show Deleted"
+            })
+        }
     } catch (err) {
         res.send({
             success: false,
@@ -70,54 +81,54 @@ router.delete('/delete-shows', authMiddleware, async (req,res)=>{
 })
 
 //get all unique theatres which have shows of a movie
-router.post('/get-all-theatres-by-movie',authMiddleware,async(req,res)=>{
-    try{
-        const {movie,date} = req.body;
+router.post('/get-all-theatres-by-movie', authMiddleware, async (req, res) => {
+    try {
+        const { movie, date } = req.body;
         //find all the shows of a movie on given date
-        const shows = await Shows.find({movie,date}).populate('theatre');
+        const shows = await Shows.find({ movie, date }).populate('theatre');
         //get all unique theatres
         let uniqueTheatre = [];
-        shows.forEach((show)=>{
+        shows.forEach((show) => {
             const isTheatre = uniqueTheatre.find(
-                (theatre)=> theatre._id == show.theatre._id
+                (theatre) => theatre._id == show.theatre._id
             )
             //Array.find returns true/false value
-            if(!isTheatre){
+            if (!isTheatre) {
                 const showsForThisTheatre = shows.filter(
-                    (showObj)=>showObj.theatre._id == show.theatre._id
+                    (showObj) => showObj.theatre._id == show.theatre._id
                 )
                 uniqueTheatre.push({
                     ...show.theatre._doc, //._doc gives all key-val pairs of the document referred to
-                    shows:showsForThisTheatre
+                    shows: showsForThisTheatre
                 });
             }
         })
         res.send({
-            success:true,
-            message:"Unique Data Fetched",
-            data:uniqueTheatre
+            success: true,
+            message: "Unique Data Fetched",
+            data: uniqueTheatre
         })
-    }catch(err){
+    } catch (err) {
         res.send({
-            success:false,
-            message:err.message
+            success: false,
+            message: err.message
         })
     }
 })
 
 
-router.post('/get-show-by-id', authMiddleware, async (req,res) => {
+router.post('/get-show-by-id', authMiddleware, async (req, res) => {
     try {
         const show = await Shows.findById(req.body.showId).populate('movie').populate('theatre');
         res.send({
-            success:true,
-            message:"Show Fetched",
-            data:show
+            success: true,
+            message: "Show Fetched",
+            data: show
         })
     } catch (error) {
         res.send({
-            success:false,
-            message:err.message
+            success: false,
+            message: err.message
         })
     }
 })
